Removes old Gaussian03 group checks on Gaussian09
We used to have an installation of Gaussian09 installed under the old
`legg03` group because of a limit on group membership and a problem
creating new groups, but both of those problems are gone now.

Note that because the `gaussian/g09-d01/pgi-2015.7` module used to
divert people to one of two installations depending on group membership,
some users will have found themselves using a different installation
than previously. This is not expected to cause any problems, as they're
the same version of Gaussian built the same way.